Cambios de permisos, cosas que no puede ver Admin

+++ src/main/java/co/com/m4h/registros/service/impl/UserServiceImpl.java
@@ -5,11 +5,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import co.com.m4h.registros.common.SecurityUtil;
 import co.com.m4h.registros.model.User;
 import co.com.m4h.registros.persistence.UserRepository;
 import co.com.m4h.registros.service.UserService;
@@ -37,8 +39,13 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public User update(User user) {
-		PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-		user.setPassword(passwordEncoder.encode(user.getPassword()));
+		user.setLastPasswordResetDate(new Date(System.currentTimeMillis()));
+		User savedUser = userRepository.findOneByUsername(user.getUsername());
+
+		if (!savedUser.getPassword().equals(user.getPassword())) {
+			PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+			user.setPassword(passwordEncoder.encode(user.getPassword()));
+		}
 
 		return userRepository.save(user);
 	}
@@ -51,9 +58,10 @@ public class UserServiceImpl implements UserService {
 	@Override
 	@Transactional
 	public void delete(Long userId) {
-		// TODO: Add security check against the company
-
-		userRepository.delete(userId);
+		if (SecurityUtil.getUserId().equals(userId))
+			throw new EmptyResultDataAccessException("No se puede eliminar el mismo usuario", 0);
+		else
+			userRepository.delete(userId);
 	}
 
 	@Override
@@ -64,8 +72,18 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public Optional<User> findByUsername(String username) {
-		// TODO: Pendiente
-		User user = userRepository.findByUsername(username);
-		return Optional.ofNullable(user);
+		String role = SecurityUtil.getRole();
+
+		if (role.equals("ROLE_ROOT")) {
+			User user = userRepository.findOneByUsername(username);
+
+			return Optional.ofNullable(user);
+		} else {
+			Long companyId = SecurityUtil.getCompanyId();
+			User user = userRepository.findOneByCompanyIdAndUsername(companyId, username);
+
+			return Optional.ofNullable(user);
+		}
+
 	}
 }
